requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
Deprecated==1.3.1
Events==0.5
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.26.0
mcp-server-opensearch==0.0.2
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.11.0
pymisp==2.5.32.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sse-starlette==3.2.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wrapt==2.1.1

# --- SOC Agent dependencies ---
# LangGraph + LangChain core
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.1.0

# MCP adapter for LangChain
langchain-mcp-adapters>=0.1.0

# Vector store
chromadb>=0.5.0
langchain-chroma>=0.1.0

# Embeddings (local, free)
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# CLI
typer>=0.12.0
rich>=13.0.0
pyyaml>=6.0
